refactor(artifacts): remove BeewareRuntime, move pip_install_command, drop _FLOORS
- Remove BeewareRuntime and its docs/dev/beeware-layout.md spike doc;
  all examples now target Python 3.15 from python.org so the interim
  fallback is no longer needed.
- Move pip_install_command from wheels.py into collect.py, the only
  caller, following the locality principle.
- Remove _FLOORS dict and ios_floor() from PythonOrgRuntime (and the
  PythonRuntime Protocol); the deployment floor is now read from the
  xcframework Info.plist by python_meta.py and carried in
  PythonXcframeworkInfo, so a second static mapping is dead code.
